fix: store 페이지 수정
